Surveillance2024_py-main/pyproject.toml
[tool.poetry]
name = "poetry-demo"
version = "0.1.0"
description = ""
authors = ["ChatwinChu <woody408234@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
opencv-python = "^4.9.0.80"
ultralytics = "^8.1.5"
imutils = "^0.5.4"
argparse = "^1.4.0"
dlib = "^19.24.2"
face-recognition = "^1.3.0"
tensorflow-io-gcs-filesystem = "0.27.0"
deepface = "^0.0.83"
tensorflow = "~2.10"
pillow = "^10.2.0"
pandas = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
